// react-type-vite/src/services/api/request/expertRequest.ts
export interface ExpertRequest {
    username: string;
    password?: string;
    email: string;
    firstName: string;
    lastName: string;
    expertId: string;
    description?: string;
    educationalUnitId?: string;
    accountStatus?: string;
    dob?: string;
    phoneNumber?: string;
    bio?: string;
}